Updated package name
Changed package name from com.example.homepage to edu.ucdenver.questingcrew.arisingquests.
