+++ nyssa/HomePage/app/src/main/java/edu/ucdenver/questingcrew/MainActivity.java
@@ -1,11 +1,11 @@
-package com.example.homepage;
+package edu.ucdenver.questingcrew;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.view.Menu;
 
-import com.google.android.material.snackbar.Snackbar;
+import edu.ucdenver.questingcrew.arisingquests.R;
 import com.google.android.material.navigation.NavigationView;
 
 import androidx.navigation.NavController;
@@ -15,7 +15,7 @@ import androidx.navigation.ui.NavigationUI;
 import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.homepage.databinding.ActivityMainBinding;
+import edu.ucdenver.questingcrew.arisingquests.databinding.ActivityMainBinding;
 
 public class MainActivity extends AppCompatActivity {
