+++ nyssa/HomePage/app/src/main/java/edu/ucdenver/questingcrew/arisingquests/ui/gallery/GalleryViewModel.java
@@ -1,16 +1,16 @@
-package com.example.homepage.ui.slideshow;
+package edu.ucdenver.questingcrew.arisingquests.ui.gallery;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class SlideshowViewModel extends ViewModel {
+public class GalleryViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public SlideshowViewModel() {
+    public GalleryViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is slideshow fragment");
+        mText.setValue("This is gallery fragment");
     }
 
     public LiveData<String> getText() {
