+++ nyssa/HomePage/app/src/main/java/edu/ucdenver/questingcrew/arisingquests/ui/home/HomeViewModel.java
@@ -1,16 +1,16 @@
-package com.example.homepage.ui.gallery;
+package edu.ucdenver.questingcrew.arisingquests.ui.home;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class GalleryViewModel extends ViewModel {
+public class HomeViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public GalleryViewModel() {
+    public HomeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is gallery fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
