+++ nyssa/HomePage/app/src/main/java/edu/ucdenver/questingcrew/arisingquests/ui/slideshow/SlideshowFragment.java
@@ -1,4 +1,4 @@
-package com.example.homepage.ui.home;
+package edu.ucdenver.questingcrew.arisingquests.ui.slideshow;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,22 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.homepage.databinding.FragmentHomeBinding;
+import edu.ucdenver.questingcrew.arisingquests.databinding.FragmentSlideshowBinding;
 
-public class HomeFragment extends Fragment {
+public class SlideshowFragment extends Fragment {
 
-    private FragmentHomeBinding binding;
+    private FragmentSlideshowBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        HomeViewModel homeViewModel =
-                new ViewModelProvider(this).get(HomeViewModel.class);
+        SlideshowViewModel slideshowViewModel =
+                new ViewModelProvider(this).get(SlideshowViewModel.class);
 
-        binding = FragmentHomeBinding.inflate(inflater, container, false);
+        binding = FragmentSlideshowBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textHome;
-        homeViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textSlideshow;
+        slideshowViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
