+++ nyssa/HomePage/app/src/main/res/layout/fragment_gallery.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.gallery.GalleryFragment">
+    tools:context="edu.ucdenver.questingcrew.arisingquests.ui.gallery.GalleryFragment">
 
     <TextView
         android:id="@+id/text_gallery"
