+++ nyssa/HomePage/app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.example.homepage.ui.home.HomeFragment"
+        android:name="edu.ucdenver.questingcrew.arisingquests.ui.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_calendar"
-        android:name="com.example.homepage.ui.gallery.GalleryFragment"
+        android:name="edu.ucdenver.questingcrew.arisingquests.ui.gallery.GalleryFragment"
         android:label="@string/menu_gallery"
         tools:layout="@layout/fragment_gallery" />
 
     <fragment
         android:id="@+id/nav_mental"
-        android:name="com.example.homepage.ui.slideshow.SlideshowFragment"
+        android:name="edu.ucdenver.questingcrew.arisingquests.ui.slideshow.SlideshowFragment"
         android:label="@string/menu_slideshow"
         tools:layout="@layout/fragment_slideshow" />
 </navigation>
